+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [UNRELEASED]
+
+### Fixed
+
+- Prevent deletion of actors during escalation.
+
 ## [2.10.2] - 2026-04-07
 
 ### Fixed